requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
curl_cffi==0.7.4
dataclasses-json==0.6.7
distro==1.9.0
docarray==0.40.0
duckduckgo_search==5.3.0
faiss-cpu==1.9.0.post1
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.6.6
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-google-genai==1.0.8
langchain-openai==0.1.7
langchain-text-splitters==0.2.4
langchainhub==0.1.21
langsmith==0.1.147
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.24.1
numpy==1.26.4
openai==1.60.2
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==10.4.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==4.25.6
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
streamlit==1.33.0
tenacity==8.5.0
tiktoken==0.8.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
yarl==1.18.3
